Removing currency spec, using generic record spec instead
I don't think the existing specs add that much value. They are a pain to write and don't tell us much.

I'm testing out a simplier "basic" record spec. If this passes, there is a good chance the NS gem logic will work fine
in the wild, although it's no gaurentee that the logic lines up with what NetSuite returns.