Combine *Addressbook* records; add EmployeeAddressbook* to Employee.
Code between the original implementation of ContactAddressbook* and CustomerAddressbook* was being duplicated, so I abstracted this away into base classes [ Addressbook and AddressbookList ]. I used these base classes with the orignal two classes, and added EmployeeAddressbook* as well.

I chose to not test the *List sublist classes as the no functionality between the base class and subclasses changed.