Clarify search row_list data components
The "records" in the search group are hashes with "keys" that are the
attribute names and "values" that the data returned by the search for
those attributes (not necessarily the attribute values, though). For
example, you'll often see things like this:

```
{
  :department=>{
    :search_value=>{:@internal_id=>"113"},
    :custom_label=>"Business Unit"
  }
}
```

I think using `attr_name` and `search_result` instead of `k` and `v`,
respectively, makes it clearer what the parts of this data constitute.