Expose non-field search results within custom field list
Sometimes NS makes information available within saved search results
that are not record fields. For example, you can run an InventoryItem
search through the GUI that returns:

* `Location Available` (the quantity available at the default location)
* `Location Reorder Point` (quantity at which you should reorder)
* `Location On Order` (quantity on order for the location)

These are not record fields, as you can see with a quick CTRL-f through
the NS schema docs:

http://www.netsuite.com/help/helpcenter/en_US/srbrowser/Browser2018_2/schema/record/inventoryitem.html

They aren't custom formula results either (which NS doesn't seem to expose
through the API).

And they also aren't _custom fields_. This information is just default,
standard-issue NetSuite. It's perfectly reasonable to have a search that
returns them as results, and thus also reasonable to expect to be able
to access them through the API.