[CLD-669-vendorcredit-record-refs]
*Why:

The vendor credit object in the gem previously had record reference fields defined differently than the majority of the other objects in the gem.  Because of this, when we call '.record_refs' in the connector, it was erroring

*This addresses the change by:

Adding in the record ref support and redefined how the record reference fields are defined.

[CLD-669]