Properly extract external_id from advanced search results
If you performed an advanced search specifying which columns to return
and included externalId, you wouldn't be able to get the external_id
from the resulting record via the usual external_id method. This now
matches how internal_id is extracted.

Prior to this, calls to external_id returned:
```
> record.external_id
=> {:@external_id=>"customer_1"}
```

Now, it returns:
```
> record.external_id
=> "customer_1"
```