Introduce search only fields
When performing a search, there's additional columns that can be
requested that aren't normally on the record. Previously such columns
were not accessible via this gem's search results.

Now these columns can be accessed just as you'd access a regular field
or read only field.

Starting with InventoryItem and Invoice.

Ultimately, these are fields of the ___SearchRowBasic,
ItemSearchRowBasic and TransactionSearchRowBasic, so eventually it may
be wise to extract a module encapulating those SearchRowBasic's and
include that in the relevant records to avoid duplication of search only
fields across multiple records (ie. Estimate, SalesOrder, Invoice,
CreditMemo, etc.) that all utilize the same SearchRowBasic.