return-netsuite-error-for-search-action-failures
Why
Currently, if the `Search` action is unsuccessful, it will simply return `false` and give no insight into what the error was.
Looking at patterns implemented elsewhere in the codebase, we seem to be moving towards returning a `NetSuite::Error` if the response is not successful.

Solution
Return a `NetSuite::Error` if the response is not successful.