Fix enum fields removed from ServiceResaleItem mistakenly
I jumped the gun in #500 and removed these fields assuming their custom
types had to be implemented in the gem, but a few were just enums, such
as `sitemap_priority`, which shouldn't need a custom type class as we
get a string value back:
```xml
<listAcct:sitemapPriority>_auto</listAcct:sitemapPriority>
```
https://system.netsuite.com/help/helpcenter/en_US/srbrowser/Browser2021_2/schema/enum/sitemappriority.html?mode=package

Perhaps one day these enums should exist in the gem so you can refer to
the values by something like a constant rather than a magic string (ie.
`SitemapPriority::AUTO` vs "_auto"), but for now, the magic string is
probably sufficient.