Update ItemVendor fields/record_refs for 2021.2
A couple were missing for my use case, so I updated them based on 2021.2

`vendor` got moved from a `field` to a `record_ref`.

I arbitrarily went back 5 years (to version 2016.1) and it was still a
`RecordRef` back then, so this doesn't appear to have been a recent
change on the NetSuite side of things that might affect users of this
gem.