Prevent fields from conflicting with other methods on records
`Records#type` conflicted with the `search_only_field :type` on
`Invoice`, such that when `AttachFile` called `@object.type`, you'd get
 the search-only field value (often `nil`), rather than the result of
 `Records#type`.

`Records#type` was renamed to `Records#netsuite_type` to avoid such
conflicts.

Furthermore, I added a check to ensure that future fields and
public/protected methods don't conflict. I tried to include private
methods too for completeness, but since we want to check up the ancestor
tree to see if a method is defined, `search_only_field :print` on
`Invoice` conflicted with `Kernel#print` in the ancestor tree. For now I
opted to turning a blind eye to conflicts with private methods, however
the alternative would be to translate to another method name for
`search_only_field :print`, similar to the `class/klass` translation.