Update customer record (update fields, add search only fields) (#535)
* Rename CustomerSubscriptionsList to SubscriptionsList and CustomerSubscription to Subscription

At least as far back as 2014.1, NetSuite used these names, without the
Customer prefix, so this aligns the gems naming with NetSuite.

This would be a breaking change for anyone using the old name.

* Update `Customer` record fields/record refs for 2021.2

The following were moved from `fields` to `record_refs`:
- buying_reason
- buying_time_frame
- campaign_category
- image
- opening_balance_account
- pref_cc_processor
- representing_subsidiary
- sales_group
- sales_readiness

The following were removed as `fields` since their sublist class is not
yet implemented:
- download_list
- group_pricing_list
- item_pricing_list

I didn't update read-only fields as I don't trust NetSuite's
documentation. They say `creditHoldOverride` is read-only, for example,
but I can very clearly edit it. Therefore opting to leave this list
as-is to avoid breaking changes.

* Add search-only fields to `Customer`