Avoid Savon version `2.13.0` to prevent generating invalid envelopes
This version of Savon mistakenly prefixed the default namespace of
`xmlns`, resulting in a namespace of `xmlns:xmlns`, which caused
NetSuite to error on requests. v2.13.1 fixed this issue to skip
prefixing the default namespace.

This is a continuation of #558, which avoided v2.13.0 and anything
greater.

Bug: https://github.com/savonrb/savon/pull/943#issuecomment-1213163418
Fix: https://github.com/savonrb/savon/pull/977