Use RC of mail gem in CI for Ruby 3.1 support
Savon v2.13 adds a dependency on mail, which has an implicit dependency
on net-smtp. Ruby 3.1 removed net-smtp as a default gem. mail v2.8.0.rc1
is the first to make that dependency explicit to support Ruby 3.1.

This at least gets CI passing, but may not help end-users on Ruby 3.1.
If they previously had a dependency on mail (ie. through Rails), there's
probably already an explicit dependency in their app on net-smtp to get
it working on Ruby 3.1. If they have no dependency on mail previously,
this will likely error as they go to require this gem, which requires
savon, which requires mail, which requires net-smtp, but Ruby can't find
it. It doesn't seem right to add net-smtp (or an RC of mail) to this
gem's gemspec dependencies, given we don't directly use either library.
Ultimately it's about the latest stable release of mail being
incompatible with Ruby 3.1.