Add search only fields to `CreditMemo` and `PurchaseOrder`
Both `Invoice`, `CreditMemo`, and `PurchaseOrder` records (as well as
many other records) use the same search fields for the returned basic
results, `TransactionSearchRowBasic`, so rather than repeat the fields
for each record, extract them to a common module that can be shared
between the respective records.

Since the search module may cover multiple records, it's fields will
overlap with some of the standard fields on each record, so defining
search only fields is now a no-op if the search only field already
exists on the record as a standard field. For this reason, it's
necessary to include the search module after defining all the standard
fields, otherwise we'll error if we try to define a standard field if
it's already been defined as a search only field.

At this point, I only have use for searching invoices, credit memos, and
purchase orders, so the module is only included on those records, but
it's the same search fields that could eventually be included on
`AssemblyBuild`, `BinTransfer`, `CashSale`, etc. records, all of which
generally seem to fall under "transaction".