Raise error message from SOAP response when initialize action fails
The current error handling approach of the initalize action is to raise
a generic failure message.

For example:
NetSuite::Records::ItemReceipt.initialize with #<NetSuite::Records::PurchaseOrder:0x00007f561d8717c8> failed.

This does not provide any of the useful context from the SOAP response
which would be more helpful for debugging the issue.

Support for fetching error context from the SOAP response XML has been
added to other actions, for example:
* https://github.com/netsweet/netsuite/commit/3b756b9d7b802375ffb9d059b303dc2e18e94d84
* https://github.com/netsweet/netsuite/commit/e6401961ef35e8d60b30382c810e893ade2e0e44
* https://github.com/netsweet/netsuite/commit/70ab07710387ce32f746a6213ec1087a0f764ae7

Adding similar support to the initialize action, and using the message
of the first error in the SOAP XML for the InitializationError raise,
would provide more useful error context for debugging.